Remove demo folder
