--- demo/graph-tutorial/package.json
@@ -1,64 +1,0 @@
-{
-  "name": "graph-tutorial",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@azure/msal-browser": "2.16.1",
-    "@azure/msal-react": "1.0.1",
-    "@microsoft/microsoft-graph-client": "3.0.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "bootstrap": "^5.0.1",
-    "date-fns": "2.23.0",
-    "date-fns-tz": "1.1.6",
-    "react": "^17.0.2",
-    "react-bootstrap": "2.0.0-beta.4",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1",
-    "windows-iana": "^5.0.2"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@types/microsoft-graph": "^2.0.0",
-    "@types/react-router-dom": "5.1.8"
-  },
-  "jest": {
-    "transformIgnorePatterns": [
-      "/node_modules/(?!date-fns)"
-    ],
-    "moduleNameMapper": {
-      "\\.(css)$": "<rootDir>/src/__mocks__/mockCss.js"
-    }
-  }
-}
