--- demo/graph-tutorial/src/Config.example.ts
@@ -1,11 +1,0 @@
-const config = {
-  appId: 'YOUR_APP_ID_HERE',
-  redirectUri: 'http://localhost:3000',
-  scopes: [
-    'user.read',
-    'mailboxsettings.read',
-    'calendars.readwrite'
-  ]
-};
-
-export default config;
