--- demo/graph-tutorial/src/__mocks__/mockCss.js
@@ -1,1 +1,0 @@
-module.exports = {};
