--- demo/graph-tutorial/src/index.css
@@ -1,9 +1,0 @@
-body {
-  padding-top: 4.5rem;
-}
-
-.alert-pre {
-  word-wrap: break-word;
-  word-break: break-all;
-  white-space: pre-wrap;
-}
