--- demo/graph-tutorial/src/index.tsx
@@ -1,54 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import {
-  PublicClientApplication,
-  EventType,
-  EventMessage,
-  AuthenticationResult } from '@azure/msal-browser';
-
-import config from './Config';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-
-// <MsalInstanceSnippet>
-const msalInstance = new PublicClientApplication({
-  auth: {
-    clientId: config.appId,
-    redirectUri: config.redirectUri
-  },
-  cache: {
-    cacheLocation: 'sessionStorage',
-    storeAuthStateInCookie: true
-  }
-});
-
-// Check if there are already accounts in the browser session
-// If so, set the first account as the active account
-const accounts = msalInstance.getAllAccounts();
-if (accounts && accounts.length > 0) {
-  msalInstance.setActiveAccount(accounts[0]);
-}
-
-msalInstance.addEventCallback((event: EventMessage) => {
-  if (event.eventType === EventType.LOGIN_SUCCESS && event.payload) {
-    // Set the active account - this simplifies token acquisition
-    const authResult = event.payload as AuthenticationResult;
-    msalInstance.setActiveAccount(authResult.account);
-  }
-});
-// </MsalInstanceSnippet>
-
-// <RenderSnippet>
-ReactDOM.render(
-  <React.StrictMode>
-    <App pca = { msalInstance }/>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
-// </RenderSnippet>
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
