--- demo/version
@@ -1,1 +1,0 @@
-1.3
